+++ XuanYu.Editor.UI/Win/UiWin.SceneCommands.cs
@@ -36,7 +36,7 @@ public partial class UiWin
     async Task RunSceneCommand(string command)
     {
         if (DataContext is not UiVm vm) return;
-        if (command is "新建地图" or "打开地图" or "保存地图" or "卸载地图" or "聚焦地图")
+        if (command is "新建地图" or "聚焦地图")
         {
             await RunMapCommand(command);
             return;